chassis: Move layer option handling into separate files
All option handling moved out of chassis.cpp into layer_options.cpp.
